third_party/blink/web_tests/editing/selection/move-by-sentence-001.html

<!DOCTYPE html>
<html> 
<body>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => {
  assert_selection('|\u0095 <b>This is some linked text</b>',
      selection => {
        for (let i = 0; i < 8; ++i) {
          selection.modify('move', 'forward', 'line');
          selection.modify('move', 'backward', 'sentence');
        }
      },
      '|\u0095 <b>This is some linked text</b>');
}, 'No crash by repeating caret movement by a sentence');

// Note that this tests the fix for this bug:
// <rdar://problem/5174862> Crash resulting from DeprecatedString::insert()
</script>
</body>
</html>
